+++ requirements.txt
@@ -1,4 +1,3 @@
 celery
-influxdb
 redis
 pytelegraf
